Create a users.serailizer package, set db_index=True on username
