+++ lego/apps/users/views/penalties.py
@@ -3,7 +3,7 @@ from rest_framework import mixins
 from rest_framework.viewsets import GenericViewSet
 
 from lego.apps.users.models import Penalty
-from lego.apps.users.serializers import PenaltySerializer
+from lego.apps.users.serializers.penalties import PenaltySerializer
 
 
 class PenaltyViewSet(mixins.CreateModelMixin,
